Move sync command to SyncCommand
