+++ s4/commands/__init__.py
@@ -1,0 +1,1 @@
+#! -*- encoding: utf8 -*-
