+++ tests/test_cli.py
@@ -12,22 +12,9 @@ import mock
 import pytest
 import pytz
 
-from s4 import cli, sync
-from s4.clients import SyncState
-from s4.resolution import Resolution
+from s4 import cli
 from s4.utils import to_timestamp
 from tests import utils
-
-
-class FakeInputStream(object):
-    def __init__(self, results):
-        self.results = results
-        self.index = 0
-
-    def __call__(self, *args, **kwargs):
-        output = self.results[self.index]
-        self.index += 1
-        return output
 
 
 @pytest.yield_fixture
@@ -53,109 +40,6 @@ def get_timestamp(year, month, day, hour, minute):
     return to_timestamp(
         datetime(year, month, day, hour, minute, tzinfo=pytz.UTC)
     )
-
-
-@mock.patch('s4.utils.get_input')
-class TestHandleConflict(object):
-    def test_first_choice(self, get_input, s3_client, local_client):
-        get_input.return_value = '1'
-
-        action_1 = SyncState(
-            SyncState.UPDATED,
-            1111, 2222
-        )
-        action_2 = SyncState(
-            SyncState.DELETED,
-            3333, 4444
-        )
-
-        result = cli.handle_conflict(
-            'movie',
-            action_1, s3_client,
-            action_2, local_client,
-        )
-        assert result.action == Resolution.UPDATE
-
-    def test_second_choice(self, get_input, s3_client, local_client):
-        get_input.return_value = '2'
-
-        action_1 = SyncState(
-            SyncState.UPDATED,
-            1111, 2222
-        )
-        action_2 = SyncState(
-            SyncState.DELETED,
-            3333, 4444
-        )
-
-        result = cli.handle_conflict(
-            'movie',
-            action_1, s3_client,
-            action_2, local_client,
-        )
-        assert result.action == Resolution.DELETE
-
-    def test_skip(self, get_input, s3_client, local_client):
-        get_input.return_value = 'X'
-
-        action_1 = SyncState(
-            SyncState.UPDATED,
-            1111, 2222
-        )
-        action_2 = SyncState(
-            SyncState.DELETED,
-            3333, 4444
-        )
-
-        result = cli.handle_conflict(
-            'movie',
-            action_1, s3_client,
-            action_2, local_client,
-        )
-        assert result is None
-
-    @mock.patch('s4.cli.show_diff')
-    def test_diff(self, show_diff, get_input, s3_client, local_client):
-        get_input.side_effect = FakeInputStream([
-            'd',
-            'X',
-        ])
-
-        action_1 = SyncState(
-            SyncState.UPDATED,
-            1111, 2222
-        )
-        action_2 = SyncState(
-            SyncState.DELETED,
-            3333, 4444
-        )
-
-        result = cli.handle_conflict(
-            'movie',
-            action_1, s3_client,
-            action_2, local_client,
-        )
-        assert result is None
-        assert show_diff.call_count == 1
-        show_diff.assert_called_with(s3_client, local_client, 'movie')
-
-
-def test_progressbar_smoketest(s3_client, local_client):
-    # Just test that nothing blows up
-    utils.set_local_contents(
-        local_client, 'history.txt',
-        data='a long long time ago',
-        timestamp=5000,
-    )
-
-    worker = sync.SyncWorker(
-        s3_client,
-        local_client,
-        start_callback=cli.display_progress_bar,
-        update_callback=cli.update_progress_bar,
-        complete_callback=cli.hide_progress_bar,
-    )
-    worker.sync()
 
 
 # TODO: Should catch KeyboardExceptions and raise them again
@@ -477,7 +361,7 @@ class TestEditCommand(object):
         )
 
     def test_no_changes(self, get_input, config_file):
-        fake_stream = FakeInputStream([
+        fake_stream = utils.FakeInputStream([
             '',
             '',
             '',
@@ -513,7 +397,7 @@ class TestEditCommand(object):
         assert expected_config == config
 
     def test_correct_output(self, get_input, config_file):
-        fake_stream = FakeInputStream([
+        fake_stream = utils.FakeInputStream([
             '/home/user/Documents',
             's3://buckets/mybackup222',
             '9999999999',
@@ -552,7 +436,7 @@ class TestEditCommand(object):
 @mock.patch('s4.utils.get_input')
 class TestAddCommand(object):
     def test_correct_behaviour(self, get_input, config_file):
-        fake_stream = FakeInputStream([
+        fake_stream = utils.FakeInputStream([
             '/home/user/Documents',
             's3://mybucket/Documents',
             'eu-west-2',
@@ -582,7 +466,7 @@ class TestAddCommand(object):
         assert new_config == expected_config
 
     def test_copy_target_credentials(self, get_input, config_file):
-        fake_stream = FakeInputStream([
+        fake_stream = utils.FakeInputStream([
             '/home/user/Animals',
             's3://mybucket/Zoo',
             'us-west-2',
@@ -625,7 +509,7 @@ class TestAddCommand(object):
         assert new_config == expected_config
 
     def test_copy_target_credentials_bad_target(self, get_input, capsys):
-        fake_stream = FakeInputStream([
+        fake_stream = utils.FakeInputStream([
             '/home/user/Animals',
             's3://mybucket/Zoo',
             'us-west-2',
@@ -644,7 +528,7 @@ class TestAddCommand(object):
         )
 
     def test_custom_target_name(self, get_input, config_file):
-        fake_stream = FakeInputStream([
+        fake_stream = utils.FakeInputStream([
             '/home/user/Music',
             's3://mybucket/Musiccccc',
             'us-west-1',
